Moved part of Registro to a Header component.
