+++ FrontEnd/src/Pages/RegistroClasesAlumno/RegistroClasesAlumno.jsx
@@ -13,9 +13,6 @@ import Snackbar from '@mui/material/Snackbar';
 import Autocomplete from '@mui/material/Autocomplete';
 import { DataGrid } from '@mui/x-data-grid';
 import Modal from '@mui/material/Modal';
-import InputLabel from '@mui/material/InputLabel';
-import FormControl from '@mui/material/FormControl';
-import Select from '@mui/material/Select';
 import SearchIcon from '@mui/icons-material/Search';
 import { useAuth0 } from '@auth0/auth0-react';
 import moment from 'moment-timezone';
@@ -38,8 +35,9 @@ import MiRegistro from '../../Components/Registro/MiRegistro';
 import Clase from '../../Components/Clase/Clase';
 import { startDateDict, endDateDict } from '../../utils/constants';
 import {
-  getNivel, getHorario, getProfesor, getCupo, calculateAge, compararFecha,
+  getNivel, getHorario, getProfesor, getCupo, compararFecha,
 } from '../../utils/utilFunctions';
+import RegistroClasesHeader from '../../Components/Registro/RegistroClasesHeader';
 
 function RegistroClasesAlumnos({ changeContent }) {
   const [items, setItems] = useState([]);
@@ -217,49 +215,6 @@ function RegistroClasesAlumnos({ changeContent }) {
     setOpenMoreInfo(!openMoreInfo);
   };
 
-  const filterClasses = async (student) => {
-    const age = calculateAge(student.fecha_de_nacimiento);
-    let waitList = [];
-    let myClasses = [];
-
-    const filter = clases.filter(
-      (clase) => Number(clase.edad_minima) < age
-        && age < (clase.edad_maxima ? Number(clase.edad_maxima) : 99),
-    ).map((aClass) => ({ ...aClass, status: '' }));
-
-    const waitListResponse = await getWaitList();
-    waitList = waitListResponse.json().filter((lista) => lista.idAlumno === student._id);
-
-    waitList.forEach((inWaitList) => {
-      const classIndex = filter.findIndex((aClass) => aClass._id === inWaitList.idClase);
-      if (classIndex !== -1) {
-        filter[classIndex].status = 'ListaEspera';
-      }
-    });
-
-    const myClassesResponse = await getClassStudent();
-    myClasses = myClassesResponse.json().filter((clase) => clase.idAlumno === student._id);
-
-    myClasses.forEach((myClass) => {
-      const classIndex = filter.findIndex((aClass) => aClass._id === myClass.idClase);
-      if (classIndex !== -1) {
-        filter[classIndex].status = 'Inscrito';
-      }
-    });
-
-    setFilteredClasses(filter);
-  };
-
-  const handleChange = (e) => {
-    if (e.target.value === '') {
-      setFilteredClasses(clases);
-      setCurrentStudent(null);
-      return;
-    }
-    setCurrentStudent(e.target.value);
-    filterClasses(e.target.value);
-  };
-
   const handleListaEspera = async (clase) => {
     const waitListResponse = await getWaitList();
     const lista = (await waitListResponse.json()).filter(
@@ -343,13 +298,6 @@ function RegistroClasesAlumnos({ changeContent }) {
     setOpenConfirmationDialog(false);
   };
 
-  const handleNameFilter = (value) => {
-    const filteredClasses = clases.filter(
-      (clase) => clase.nombre_curso.toLowerCase().includes(value.trim().toLowerCase()),
-    );
-    setFilteredClasses(filteredClasses);
-  };
-
   if (!students || !clases) {
     return (
       <Box
@@ -397,47 +345,14 @@ function RegistroClasesAlumnos({ changeContent }) {
 
   return (
     <Box>
-      <Typography variant="h3" sx={{ m: 2, color: '#004a98' }}>
-        Registro clases (Inscripción)
-      </Typography>
-      <Box sx={{ m: 2, position: 'sticky', top: '10px' }}>
-        <FormControl fullWidth>
-          <InputLabel>Estudiantes</InputLabel>
-          <Select
-            value={currentStudent || ''}
-            label="Estudiantes"
-            onChange={handleChange}
-          >
-            <MenuItem value="">
-              <em>Estudiante</em>
-            </MenuItem>
-            {students.map((student) => (
-              <MenuItem key={student._id} value={student}>
-                {student.nombre}
-                {' '}
-                {student.apellido_paterno}
-                {' '}
-                {student.apellido_materno}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-        <Autocomplete
-          disablePortal
-          options={classNames}
-          onChange={(e, newvalue) => handleNameFilter(newvalue)}
-          onInputChange={(e, newvalue) => handleNameFilter(newvalue)}
-          sx={{ display: { xs: 'flex', md: 'none' }, mt: 1 }}
-          fullWidth
-          renderInput={(params) => (
-            <TextField
-              {...params}
-              label="Curso"
-              helperText="Busca tu curso"
-            />
-          )}
-        />
-      </Box>
+      <RegistroClasesHeader 
+        classNames={classNames}
+        clases={clases}
+        currentStudent={currentStudent}
+        students={students}
+        setFilteredClasses={setFilteredClasses}
+        setCurrentStudent={setCurrentStudent}
+      />
       <Box
         sx={{
           textAlign: 'center',
